Remove unused widget test file
The `widget_test.dart` file in the `front_flutter/test` directory was deleted. This file contained a basic Flutter widget test that was no longer needed. The test was checking the increment functionality of a counter widget, but it appears that this test is now redundant or obsolete. Removing this file helps to clean up the codebase and reduce clutter.